consolidate verify results filename in nameUtils
